Refactoring Evse15118D20: Move some anoynm functions in utils.hpp/cpp
