+++ modules/EVSE/Evse15118D20/charger/ISO15118_chargerImpl.cpp
@@ -23,26 +23,15 @@ iso15118::config::TlsNegotiationStrategy convert_tls_negotiation_strategy(const 
     using Strategy = iso15118::config::TlsNegotiationStrategy;
     if (strategy == "ACCEPT_CLIENT_OFFER") {
         return Strategy::ACCEPT_CLIENT_OFFER;
-    } else if (strategy == "ENFORCE_TLS") {
-        return Strategy::ENFORCE_TLS;
-    } else if (strategy == "ENFORCE_NO_TLS") {
-        return Strategy::ENFORCE_NO_TLS;
-    } else {
-        EVLOG_AND_THROW(Everest::EverestConfigError("Invalid choice for tls_negotiation_strategy: " + strategy));
-        // better safe than sorry
     }
-}
-
-template <typename T, typename U> std::optional<T> convert_from_optional(const std::optional<U>& in) {
-    return (in.has_value()) ? std::make_optional(static_cast<T>(*in)) : std::nullopt;
-}
-
-template <> std::optional<float> convert_from_optional(const std::optional<dt::RationalNumber>& in) {
-    return (in.has_value()) ? std::make_optional(dt::from_RationalNumber(*in)) : std::nullopt;
-}
-
-template <> std::optional<dt::RationalNumber> convert_from_optional(const std::optional<float>& in) {
-    return (in.has_value()) ? std::make_optional(dt::from_float(*in)) : std::nullopt;
+    if (strategy == "ENFORCE_TLS") {
+        return Strategy::ENFORCE_TLS;
+    }
+    if (strategy == "ENFORCE_NO_TLS") {
+        return Strategy::ENFORCE_NO_TLS;
+    }
+    EVLOG_AND_THROW(Everest::EverestConfigError("Invalid choice for tls_negotiation_strategy: " + strategy));
+    // better safe than sorry
 }
 
 types::iso15118::DisplayParameters convert_display_parameters(const dt::DisplayParameters& in) {
@@ -56,87 +45,6 @@ types::iso15118::DisplayParameters convert_display_parameters(const dt::DisplayP
             in.charging_complete,
             convert_from_optional<float>(in.battery_energy_capacity),
             in.inlet_hot};
-}
-
-types::iso15118::DcChargeDynamicModeValues convert_dynamic_values(const dt::Dynamic_DC_CLReqControlMode& in) {
-    return {dt::from_RationalNumber(in.target_energy_request),
-            dt::from_RationalNumber(in.max_energy_request),
-            dt::from_RationalNumber(in.min_energy_request),
-            dt::from_RationalNumber(in.max_charge_power),
-            dt::from_RationalNumber(in.min_charge_power),
-            dt::from_RationalNumber(in.max_charge_current),
-            dt::from_RationalNumber(in.max_voltage),
-            dt::from_RationalNumber(in.min_voltage),
-            convert_from_optional<float>(in.departure_time),
-            std::nullopt,
-            std::nullopt,
-            std::nullopt,
-            std::nullopt,
-            std::nullopt};
-}
-
-types::iso15118::DcChargeDynamicModeValues
-convert_dynamic_values(const iso15118::message_20::datatypes::BPT_Dynamic_DC_CLReqControlMode& in) {
-    return {dt::from_RationalNumber(in.target_energy_request),
-            dt::from_RationalNumber(in.max_energy_request),
-            dt::from_RationalNumber(in.min_energy_request),
-            dt::from_RationalNumber(in.max_charge_power),
-            dt::from_RationalNumber(in.min_charge_power),
-            dt::from_RationalNumber(in.max_charge_current),
-            dt::from_RationalNumber(in.max_voltage),
-            dt::from_RationalNumber(in.min_voltage),
-            convert_from_optional<float>(in.departure_time),
-            dt::from_RationalNumber(in.max_discharge_power),
-            dt::from_RationalNumber(in.min_discharge_power),
-            dt::from_RationalNumber(in.max_discharge_current),
-            convert_from_optional<float>(in.max_v2x_energy_request),
-            convert_from_optional<float>(in.min_v2x_energy_request)};
-}
-
-dt::Parameter convert_parameter(const types::iso15118_vas::Parameter& parameter) {
-    dt::Parameter out;
-    out.name = parameter.name;
-
-    if (parameter.value.bool_value.has_value()) {
-        out.value = parameter.value.bool_value.value();
-    } else if (parameter.value.int_value.has_value()) {
-        out.value = (int32_t)parameter.value.int_value.value();
-    } else if (parameter.value.short_value.has_value()) {
-        out.value = (int16_t)parameter.value.short_value.value();
-    } else if (parameter.value.byte_value.has_value()) {
-        out.value = (int8_t)parameter.value.byte_value.value();
-    } else if (parameter.value.rational_number.has_value()) {
-        out.value = dt::from_float(parameter.value.rational_number.value());
-    } else if (parameter.value.finite_string.has_value()) {
-        out.value = parameter.value.finite_string.value();
-    } else {
-        throw std::invalid_argument("Invalid ParameterValue in convert_parameter: " + parameter.name +
-                                    " has no value set");
-    }
-
-    return out;
-}
-
-dt::ParameterSet convert_parameter_set(const types::iso15118_vas::ParameterSet& parameter_set) {
-    dt::ParameterSet out;
-    out.id = parameter_set.set_id;
-
-    out.parameter.reserve(parameter_set.parameters.size());
-    for (const auto& parameter : parameter_set.parameters) {
-        out.parameter.push_back(convert_parameter(parameter));
-    }
-
-    return out;
-}
-
-std::vector<dt::ParameterSet>
-convert_parameter_set_list(const std::vector<types::iso15118_vas::ParameterSet>& parameter_set_list) {
-    std::vector<dt::ParameterSet> out;
-    out.reserve(parameter_set_list.size());
-    for (const auto& parameter_set : parameter_set_list) {
-        out.push_back(convert_parameter_set(parameter_set));
-    }
-    return out;
 }
 
 auto fill_mobility_needs_modes_from_config(const module::Conf& module_config) {
@@ -162,388 +70,6 @@ auto fill_mobility_needs_modes_from_config(const module::Conf& module_config) {
     return mobility_needs_modes;
 }
 
-template <typename EVSE, typename EV>
-void fill_v2x_charging_parameters(types::iso15118::V2XChargingParameters& out_params, const EVSE& evse_limits,
-                                  const EV& ev_limits);
-
-template <typename In>
-void fill_v2x_charging_parameters(types::iso15118::V2XChargingParameters& out_params, const In& data);
-
-template <>
-void fill_v2x_charging_parameters(types::iso15118::V2XChargingParameters& out_params,
-                                  const iso15118::d20::DcTransferLimits& evse_limits,
-                                  const dt::DC_CPDReqEnergyTransferMode& ev_limits) {
-    // As per the OCPP 2.1 spec (2.109) we should use the MIN/MAX function between EV and EVSE
-    out_params.min_charge_power = std::max(dt::from_RationalNumber(evse_limits.charge_limits.power.min),
-                                           dt::from_RationalNumber(ev_limits.min_charge_power));
-    out_params.max_charge_power = std::min(dt::from_RationalNumber(evse_limits.charge_limits.power.max),
-                                           dt::from_RationalNumber(ev_limits.max_charge_power));
-
-    out_params.min_charge_current = std::max(dt::from_RationalNumber(evse_limits.charge_limits.current.min),
-                                             dt::from_RationalNumber(ev_limits.min_charge_current));
-    out_params.max_charge_current = std::min(dt::from_RationalNumber(evse_limits.charge_limits.current.max),
-                                             dt::from_RationalNumber(ev_limits.max_charge_current));
-
-    out_params.min_voltage =
-        std::max(dt::from_RationalNumber(evse_limits.voltage.min), dt::from_RationalNumber(ev_limits.min_voltage));
-    out_params.max_voltage =
-        std::min(dt::from_RationalNumber(evse_limits.voltage.max), dt::from_RationalNumber(ev_limits.max_voltage));
-
-    out_params.target_soc = ev_limits.target_soc;
-}
-
-template <>
-void fill_v2x_charging_parameters(types::iso15118::V2XChargingParameters& out_params,
-                                  const iso15118::d20::DcTransferLimits& evse_limits,
-                                  const dt::BPT_DC_CPDReqEnergyTransferMode& ev_limits) {
-    // Fill in the common data
-    fill_v2x_charging_parameters<iso15118::d20::DcTransferLimits, dt::DC_CPDReqEnergyTransferMode>(
-        out_params, evse_limits, static_cast<const dt::DC_CPDReqEnergyTransferMode&>(ev_limits));
-
-    // Fill in the bidi data
-    if (evse_limits.discharge_limits.has_value()) {
-        const auto& evse_discharge_limits = evse_limits.discharge_limits.value();
-
-        out_params.min_discharge_power = std::max(dt::from_RationalNumber(evse_discharge_limits.power.min),
-                                                  dt::from_RationalNumber(ev_limits.min_discharge_power));
-        out_params.max_discharge_power = std::min(dt::from_RationalNumber(evse_discharge_limits.power.max),
-                                                  dt::from_RationalNumber(ev_limits.max_discharge_power));
-
-        out_params.min_discharge_current = std::max(dt::from_RationalNumber(evse_discharge_limits.current.min),
-                                                    dt::from_RationalNumber(ev_limits.min_discharge_current));
-        out_params.max_discharge_current = std::min(dt::from_RationalNumber(evse_discharge_limits.current.max),
-                                                    dt::from_RationalNumber(ev_limits.max_discharge_current));
-    }
-}
-
-template <>
-void fill_v2x_charging_parameters(types::iso15118::V2XChargingParameters& out_params,
-                                  const iso15118::d20::AcTransferLimits& evse_limits,
-                                  const dt::AC_CPDReqEnergyTransferMode& ev_limits) {
-    // As per the OCPP 2.1 spec (2.109) we should use the MIN/MAX function between EV and EVSE
-    out_params.min_charge_power = std::max(dt::from_RationalNumber(evse_limits.charge_power.min),
-                                           dt::from_RationalNumber(ev_limits.min_charge_power));
-    out_params.max_charge_power = std::min(dt::from_RationalNumber(evse_limits.charge_power.max),
-                                           dt::from_RationalNumber(ev_limits.max_charge_power));
-
-    if (evse_limits.charge_power_L2.has_value() and ev_limits.max_charge_power_L2.has_value() and
-        ev_limits.min_charge_power_L2.has_value()) {
-        out_params.min_charge_power_l2 = std::max(dt::from_RationalNumber(evse_limits.charge_power_L2.value().min),
-                                                  dt::from_RationalNumber(ev_limits.min_charge_power_L2.value()));
-        out_params.max_charge_power_l2 = std::min(dt::from_RationalNumber(evse_limits.charge_power_L2.value().max),
-                                                  dt::from_RationalNumber(ev_limits.max_charge_power_L2.value()));
-    }
-
-    if (evse_limits.charge_power_L3.has_value() and ev_limits.max_charge_power_L3.has_value() and
-        ev_limits.min_charge_power_L3.has_value()) {
-        out_params.min_charge_power_l3 = std::max(dt::from_RationalNumber(evse_limits.charge_power_L3.value().min),
-                                                  dt::from_RationalNumber(ev_limits.min_charge_power_L3.value()));
-        out_params.max_charge_power_l3 = std::min(dt::from_RationalNumber(evse_limits.charge_power_L3.value().max),
-                                                  dt::from_RationalNumber(ev_limits.max_charge_power_L3.value()));
-    }
-}
-
-template <>
-void fill_v2x_charging_parameters(types::iso15118::V2XChargingParameters& out_params,
-                                  const iso15118::d20::AcTransferLimits& evse_limits,
-                                  const dt::BPT_AC_CPDReqEnergyTransferMode& ev_limits) {
-    // Fill in the common data
-    fill_v2x_charging_parameters<iso15118::d20::AcTransferLimits, dt::AC_CPDReqEnergyTransferMode>(
-        out_params, evse_limits, static_cast<const dt::AC_CPDReqEnergyTransferMode&>(ev_limits));
-
-    if (evse_limits.discharge_power.has_value()) {
-        const auto& evse_discharge_limits = evse_limits.discharge_power.value();
-
-        out_params.min_discharge_power = std::max(dt::from_RationalNumber(evse_discharge_limits.min),
-                                                  dt::from_RationalNumber(ev_limits.min_discharge_power));
-        out_params.max_discharge_power = std::min(dt::from_RationalNumber(evse_discharge_limits.max),
-                                                  dt::from_RationalNumber(ev_limits.max_discharge_power));
-    }
-
-    if (evse_limits.discharge_power_L2.has_value() && ev_limits.min_discharge_power_L2.has_value() &&
-        ev_limits.max_discharge_power_L2.has_value()) {
-        const auto& evse_discharge_limits = evse_limits.discharge_power_L2.value();
-
-        out_params.min_discharge_power_l2 = std::max(dt::from_RationalNumber(evse_discharge_limits.min),
-                                                     dt::from_RationalNumber(ev_limits.min_discharge_power_L2.value()));
-        out_params.max_discharge_power_l2 = std::min(dt::from_RationalNumber(evse_discharge_limits.max),
-                                                     dt::from_RationalNumber(ev_limits.max_discharge_power_L2.value()));
-    }
-
-    if (evse_limits.discharge_power_L3.has_value() && ev_limits.min_discharge_power_L3.has_value() &&
-        ev_limits.max_discharge_power_L3.has_value()) {
-        const auto& evse_discharge_limits = evse_limits.discharge_power_L3.value();
-
-        out_params.min_discharge_power_l3 = std::max(dt::from_RationalNumber(evse_discharge_limits.min),
-                                                     dt::from_RationalNumber(ev_limits.min_discharge_power_L3.value()));
-        out_params.max_discharge_power_l3 = std::min(dt::from_RationalNumber(evse_discharge_limits.max),
-                                                     dt::from_RationalNumber(ev_limits.max_discharge_power_L3.value()));
-    }
-}
-
-template <>
-void fill_v2x_charging_parameters(types::iso15118::V2XChargingParameters& out_params,
-                                  const dt::Scheduled_SEReqControlMode& ev_control_mode) {
-    out_params.ev_target_energy_request = convert_from_optional<float>(ev_control_mode.target_energy);
-    out_params.ev_min_energy_request = convert_from_optional<float>(ev_control_mode.min_energy);
-    out_params.ev_max_energy_request = convert_from_optional<float>(ev_control_mode.max_energy);
-}
-
-template <>
-void fill_v2x_charging_parameters(types::iso15118::V2XChargingParameters& out_params,
-                                  const dt::Dynamic_SEReqControlMode& ev_control_mode) {
-    out_params.ev_target_energy_request = dt::from_RationalNumber(ev_control_mode.target_energy);
-    out_params.ev_min_energy_request = dt::from_RationalNumber(ev_control_mode.min_energy);
-    out_params.ev_max_energy_request = dt::from_RationalNumber(ev_control_mode.max_energy);
-
-    out_params.ev_min_v2xenergy_request = convert_from_optional<float>(ev_control_mode.min_v2x_energy);
-    out_params.ev_max_v2xenergy_request = convert_from_optional<float>(ev_control_mode.max_v2x_energy);
-}
-
-template <typename T> types::iso15118::AcEvPowerLimits fill_ac_ev_power_limits(const T& mode);
-
-template <> types::iso15118::AcEvPowerLimits fill_ac_ev_power_limits(const dt::AC_CPDReqEnergyTransferMode& mode) {
-    const auto max_charge_L1 = dt::from_RationalNumber(mode.max_charge_power);
-    const auto max_charge_L2 = convert_from_optional<float>(mode.max_charge_power_L2);
-    const auto max_charge_L3 = convert_from_optional<float>(mode.max_charge_power_L3);
-    const auto max_charge_total = max_charge_L1 + max_charge_L2.value_or(0.0) + max_charge_L3.value_or(0.0);
-
-    const auto min_charge_L1 = dt::from_RationalNumber(mode.min_charge_power);
-    const auto min_charge_L2 = convert_from_optional<float>(mode.min_charge_power_L2);
-    const auto min_charge_L3 = convert_from_optional<float>(mode.min_charge_power_L3);
-    const auto min_charge_total = min_charge_L1 + min_charge_L2.value_or(0.0) + min_charge_L3.value_or(0.0);
-
-    types::iso15118::AcEvPowerLimits ac_ev_power_limits;
-
-    ac_ev_power_limits.max_charge_power = {max_charge_total, std::make_optional<float>(max_charge_L1), max_charge_L2,
-                                           max_charge_L3};
-    ac_ev_power_limits.min_charge_power = {min_charge_total, std::make_optional<float>(min_charge_L1), min_charge_L2,
-                                           min_charge_L3};
-
-    return ac_ev_power_limits;
-}
-
-template <> types::iso15118::AcEvPowerLimits fill_ac_ev_power_limits(const dt::BPT_AC_CPDReqEnergyTransferMode& mode) {
-
-    const auto max_charge_L1 = dt::from_RationalNumber(mode.max_charge_power);
-    const auto max_charge_L2 = convert_from_optional<float>(mode.max_charge_power_L2);
-    const auto max_charge_L3 = convert_from_optional<float>(mode.max_charge_power_L3);
-    const auto max_charge_total = max_charge_L1 + max_charge_L2.value_or(0.0) + max_charge_L3.value_or(0.0);
-
-    const auto min_charge_L1 = dt::from_RationalNumber(mode.min_charge_power);
-    const auto min_charge_L2 = convert_from_optional<float>(mode.min_charge_power_L2);
-    const auto min_charge_L3 = convert_from_optional<float>(mode.min_charge_power_L3);
-    const auto min_charge_total = min_charge_L1 + min_charge_L2.value_or(0.0) + min_charge_L3.value_or(0.0);
-
-    const auto max_discharge_L1 = dt::from_RationalNumber(mode.max_discharge_power);
-    const auto max_discharge_L2 = convert_from_optional<float>(mode.max_discharge_power_L2);
-    const auto max_discharge_L3 = convert_from_optional<float>(mode.max_discharge_power_L3);
-    const auto max_discharge_total = max_discharge_L1 + max_discharge_L2.value_or(0.0) + max_discharge_L3.value_or(0.0);
-
-    const auto min_discharge_L1 = dt::from_RationalNumber(mode.min_discharge_power);
-    const auto min_discharge_L2 = convert_from_optional<float>(mode.min_discharge_power_L2);
-    const auto min_discharge_L3 = convert_from_optional<float>(mode.min_discharge_power_L3);
-    const auto min_discharge_total = min_discharge_L1 + min_discharge_L2.value_or(0.0) + min_discharge_L3.value_or(0.0);
-
-    types::iso15118::AcEvPowerLimits ac_ev_power_limits;
-
-    ac_ev_power_limits.max_charge_power = {max_charge_total, std::make_optional<float>(max_charge_L1), max_charge_L2,
-                                           max_charge_L3};
-    ac_ev_power_limits.min_charge_power = {min_charge_total, std::make_optional<float>(min_charge_L1), min_charge_L2,
-                                           min_charge_L3};
-    ac_ev_power_limits.max_discharge_power = {max_discharge_total, std::make_optional<float>(max_discharge_L1),
-                                              max_discharge_L2, max_discharge_L3};
-    ac_ev_power_limits.min_discharge_power = {min_discharge_total, std::make_optional<float>(min_discharge_L1),
-                                              min_discharge_L2, min_discharge_L3};
-
-    return ac_ev_power_limits;
-}
-
-template <> types::iso15118::AcEvPowerLimits fill_ac_ev_power_limits(const dt::Dynamic_AC_CLReqControlMode& mode) {
-    const auto max_charge_L1 = dt::from_RationalNumber(mode.max_charge_power);
-    const auto max_charge_L2 = convert_from_optional<float>(mode.max_charge_power_L2);
-    const auto max_charge_L3 = convert_from_optional<float>(mode.max_charge_power_L3);
-    const auto max_charge_total = max_charge_L1 + max_charge_L2.value_or(0.0) + max_charge_L3.value_or(0.0);
-
-    const auto min_charge_L1 = dt::from_RationalNumber(mode.min_charge_power);
-    const auto min_charge_L2 = convert_from_optional<float>(mode.min_charge_power_L2);
-    const auto min_charge_L3 = convert_from_optional<float>(mode.min_charge_power_L3);
-    const auto min_charge_total = min_charge_L1 + min_charge_L2.value_or(0.0) + min_charge_L3.value_or(0.0);
-
-    types::iso15118::AcEvPowerLimits ac_ev_power_limits;
-
-    ac_ev_power_limits.max_charge_power = {max_charge_total, std::make_optional<float>(max_charge_L1), max_charge_L2,
-                                           max_charge_L3};
-    ac_ev_power_limits.min_charge_power = {min_charge_total, std::make_optional<float>(min_charge_L1), min_charge_L2,
-                                           min_charge_L3};
-
-    return ac_ev_power_limits;
-}
-
-template <> types::iso15118::AcEvPowerLimits fill_ac_ev_power_limits(const dt::BPT_Dynamic_AC_CLReqControlMode& mode) {
-
-    const auto max_charge_L1 = dt::from_RationalNumber(mode.max_charge_power);
-    const auto max_charge_L2 = convert_from_optional<float>(mode.max_charge_power_L2);
-    const auto max_charge_L3 = convert_from_optional<float>(mode.max_charge_power_L3);
-    const auto max_charge_total = max_charge_L1 + max_charge_L2.value_or(0.0) + max_charge_L3.value_or(0.0);
-
-    const auto min_charge_L1 = dt::from_RationalNumber(mode.min_charge_power);
-    const auto min_charge_L2 = convert_from_optional<float>(mode.min_charge_power_L2);
-    const auto min_charge_L3 = convert_from_optional<float>(mode.min_charge_power_L3);
-    const auto min_charge_total = min_charge_L1 + min_charge_L2.value_or(0.0) + min_charge_L3.value_or(0.0);
-
-    const auto max_discharge_L1 = dt::from_RationalNumber(mode.max_discharge_power);
-    const auto max_discharge_L2 = convert_from_optional<float>(mode.max_discharge_power_L2);
-    const auto max_discharge_L3 = convert_from_optional<float>(mode.max_discharge_power_L3);
-    const auto max_discharge_total = max_discharge_L1 + max_discharge_L2.value_or(0.0) + max_discharge_L3.value_or(0.0);
-
-    const auto min_discharge_L1 = dt::from_RationalNumber(mode.min_discharge_power);
-    const auto min_discharge_L2 = convert_from_optional<float>(mode.min_discharge_power_L2);
-    const auto min_discharge_L3 = convert_from_optional<float>(mode.min_discharge_power_L3);
-    const auto min_discharge_total = min_discharge_L1 + min_discharge_L2.value_or(0.0) + min_discharge_L3.value_or(0.0);
-
-    types::iso15118::AcEvPowerLimits ac_ev_power_limits;
-
-    ac_ev_power_limits.max_charge_power = {max_charge_total, std::make_optional<float>(max_charge_L1), max_charge_L2,
-                                           max_charge_L3};
-    ac_ev_power_limits.min_charge_power = {min_charge_total, std::make_optional<float>(min_charge_L1), min_charge_L2,
-                                           min_charge_L3};
-    ac_ev_power_limits.max_discharge_power = {max_discharge_total, std::make_optional<float>(max_discharge_L1),
-                                              max_discharge_L2, max_discharge_L3};
-    ac_ev_power_limits.min_discharge_power = {min_discharge_total, std::make_optional<float>(min_discharge_L1),
-                                              min_discharge_L2, min_discharge_L3};
-
-    return ac_ev_power_limits;
-}
-
-template <> types::iso15118::AcEvPowerLimits fill_ac_ev_power_limits(const dt::Scheduled_AC_CLReqControlMode& mode) {
-    const auto max_charge_L1 = convert_from_optional<float>(mode.max_charge_power);
-    const auto max_charge_L2 = convert_from_optional<float>(mode.max_charge_power_L2);
-    const auto max_charge_L3 = convert_from_optional<float>(mode.max_charge_power_L3);
-    const auto max_charge_total =
-        max_charge_L1.value_or(0.0) + max_charge_L2.value_or(0.0) + max_charge_L3.value_or(0.0);
-
-    const auto min_charge_L1 = convert_from_optional<float>(mode.min_charge_power);
-    const auto min_charge_L2 = convert_from_optional<float>(mode.min_charge_power_L2);
-    const auto min_charge_L3 = convert_from_optional<float>(mode.min_charge_power_L3);
-    const auto min_charge_total =
-        min_charge_L1.value_or(0.0) + min_charge_L2.value_or(0.0) + min_charge_L3.value_or(0.0);
-
-    types::iso15118::AcEvPowerLimits ac_ev_power_limits;
-
-    ac_ev_power_limits.max_charge_power =
-        (max_charge_L1.has_value() or max_charge_L2.has_value() or max_charge_L3.has_value())
-            ? std::make_optional<types::units::Power>({max_charge_total, max_charge_L1, max_charge_L2, max_charge_L3})
-            : std::nullopt;
-    ac_ev_power_limits.min_charge_power =
-        (max_charge_L1.has_value() or max_charge_L2.has_value() or max_charge_L3.has_value())
-            ? std::make_optional<types::units::Power>({min_charge_total, min_charge_L1, min_charge_L2, min_charge_L3})
-            : std::nullopt;
-
-    return ac_ev_power_limits;
-}
-
-template <>
-types::iso15118::AcEvPowerLimits fill_ac_ev_power_limits(const dt::BPT_Scheduled_AC_CLReqControlMode& mode) {
-    const auto max_charge_L1 = convert_from_optional<float>(mode.max_charge_power);
-    const auto max_charge_L2 = convert_from_optional<float>(mode.max_charge_power_L2);
-    const auto max_charge_L3 = convert_from_optional<float>(mode.max_charge_power_L3);
-    const auto max_charge_total =
-        max_charge_L1.value_or(0.0) + max_charge_L2.value_or(0.0) + max_charge_L3.value_or(0.0);
-
-    const auto min_charge_L1 = convert_from_optional<float>(mode.min_charge_power);
-    const auto min_charge_L2 = convert_from_optional<float>(mode.min_charge_power_L2);
-    const auto min_charge_L3 = convert_from_optional<float>(mode.min_charge_power_L3);
-    const auto min_charge_total =
-        min_charge_L1.value_or(0.0) + min_charge_L2.value_or(0.0) + min_charge_L3.value_or(0.0);
-
-    const auto max_discharge_L1 = convert_from_optional<float>(mode.max_discharge_power);
-    const auto max_discharge_L2 = convert_from_optional<float>(mode.max_discharge_power_L2);
-    const auto max_discharge_L3 = convert_from_optional<float>(mode.max_discharge_power_L3);
-    const auto max_discharge_total =
-        max_discharge_L1.value_or(0.0) + max_discharge_L2.value_or(0.0) + max_discharge_L3.value_or(0.0);
-
-    const auto min_discharge_L1 = convert_from_optional<float>(mode.min_discharge_power);
-    const auto min_discharge_L2 = convert_from_optional<float>(mode.min_discharge_power_L2);
-    const auto min_discharge_L3 = convert_from_optional<float>(mode.min_discharge_power_L3);
-    const auto min_discharge_total =
-        min_discharge_L1.value_or(0.0) + min_discharge_L2.value_or(0.0) + min_discharge_L3.value_or(0.0);
-
-    types::iso15118::AcEvPowerLimits ac_ev_power_limits;
-
-    ac_ev_power_limits.max_charge_power =
-        (max_charge_L1.has_value() or max_charge_L2.has_value() or max_charge_L3.has_value())
-            ? std::make_optional<types::units::Power>({max_charge_total, max_charge_L1, max_charge_L2, max_charge_L3})
-            : std::nullopt;
-    ac_ev_power_limits.min_charge_power =
-        (min_charge_L1.has_value() or min_charge_L2.has_value() or min_charge_L3.has_value())
-            ? std::make_optional<types::units::Power>({min_charge_total, min_charge_L1, min_charge_L2, min_charge_L3})
-            : std::nullopt;
-    ac_ev_power_limits.max_discharge_power =
-        (max_discharge_L1.has_value() or max_discharge_L2.has_value() or max_discharge_L3.has_value())
-            ? std::make_optional<types::units::Power>(
-                  {max_discharge_total, max_discharge_L1, max_discharge_L2, max_discharge_L3})
-            : std::nullopt;
-    ac_ev_power_limits.min_discharge_power =
-        (min_discharge_L1.has_value() or min_discharge_L2.has_value() or min_discharge_L3.has_value())
-            ? std::make_optional<types::units::Power>(
-                  {min_discharge_total, min_discharge_L1, min_discharge_L2, min_discharge_L3})
-            : std::nullopt;
-    return ac_ev_power_limits;
-}
-types::iso15118::AcEvPresentPowerValues fill_ac_ev_present_power_values(const dt::Dynamic_AC_CLReqControlMode& mode) {
-    types::iso15118::AcEvPresentPowerValues present_values{};
-
-    const auto present_active_power_L1 = dt::from_RationalNumber(mode.present_active_power);
-    const auto present_active_power_L2 = convert_from_optional<float>(mode.present_active_power_L2);
-    const auto present_active_power_L3 = convert_from_optional<float>(mode.present_active_power_L3);
-    const auto present_active_power_total =
-        present_active_power_L1 + present_active_power_L2.value_or(0.f) + present_active_power_L3.value_or(0.f);
-
-    const auto present_reactive_power_L1 = dt::from_RationalNumber(mode.present_reactive_power);
-    const auto present_reactive_power_L2 = convert_from_optional<float>(mode.present_reactive_power_L2);
-    const auto present_reactive_power_L3 = convert_from_optional<float>(mode.present_reactive_power_L3);
-    const auto present_reactive_power_total =
-        present_reactive_power_L1 + present_reactive_power_L2.value_or(0.f) + present_reactive_power_L3.value_or(0.f);
-
-    present_values.present_active_power = {present_active_power_total, present_active_power_L1, present_active_power_L2,
-                                           present_active_power_L3};
-
-    present_values.present_reactive_power = {present_reactive_power_total, present_reactive_power_L1,
-                                             present_reactive_power_L2, present_reactive_power_L3};
-
-    return present_values;
-}
-
-types::iso15118::AcEvPresentPowerValues fill_ac_ev_present_power_values(const dt::Scheduled_AC_CLReqControlMode& mode) {
-    types::iso15118::AcEvPresentPowerValues present_values{};
-
-    const auto present_active_power_L1 = dt::from_RationalNumber(mode.present_active_power);
-    const auto present_active_power_L2 = convert_from_optional<float>(mode.present_active_power_L2);
-    const auto present_active_power_L3 = convert_from_optional<float>(mode.present_active_power_L3);
-    const auto present_active_power_total =
-        present_active_power_L1 + present_active_power_L2.value_or(0.f) + present_active_power_L3.value_or(0.f);
-
-    const auto present_reactive_power_L1 = convert_from_optional<float>(mode.present_reactive_power);
-    const auto present_reactive_power_L2 = convert_from_optional<float>(mode.present_reactive_power_L2);
-    const auto present_reactive_power_L3 = convert_from_optional<float>(mode.present_reactive_power_L3);
-    const auto present_reactive_power_total = present_reactive_power_L1.value_or(0.f) +
-                                              present_reactive_power_L2.value_or(0.f) +
-                                              present_reactive_power_L3.value_or(0.f);
-
-    present_values.present_active_power = {present_active_power_total, present_active_power_L1, present_active_power_L2,
-                                           present_active_power_L3};
-
-    present_values.present_reactive_power =
-        (present_reactive_power_L1.has_value() or present_reactive_power_L2.has_value() or
-         present_reactive_power_L3.has_value())
-            ? std::make_optional<types::units::Power>({present_reactive_power_total, present_reactive_power_L1,
-                                                       present_reactive_power_L2, present_reactive_power_L3})
-            : std::nullopt;
-
-    return present_values;
-}
-
 types::iso15118::AcEvDynamicModeValues fill_ac_ev_dynamic_control_mode(const dt::Dynamic_AC_CLReqControlMode& mode) {
     types::iso15118::AcEvDynamicModeValues values{};
 
@@ -557,7 +83,7 @@ types::iso15118::AcEvDynamicModeValues fill_ac_ev_dynamic_control_mode(const dt:
 
 types::iso15118::EnergyTransferMode get_energy_transfer_mode(const dt::ServiceCategory& service_category,
                                                              const std::optional<dt::AcConnector>& ac_connector) {
-    using namespace types::iso15118;
+    using EnergyTransferMode = types::iso15118::EnergyTransferMode;
 
     EnergyTransferMode requested_energy_transfer = EnergyTransferMode::AC_single_phase_core;
 
@@ -770,9 +296,8 @@ iso15118::session::feedback::Callbacks ISO15118_chargerImpl::create_callbacks() 
                const feedback::EvSEControlMode& ev_control_mode,
                const std::vector<dt::ServiceCategory>& ev_energy_services) {
             // Everest types sent to OCPP
-            using namespace types::iso15118;
 
-            ChargingNeeds charging_needs;
+            types::iso15118::ChargingNeeds charging_needs;
 
             charging_needs.requested_energy_transfer = get_energy_transfer_mode(service_category, ac_connector);
             if (!ev_energy_services.empty()) {
@@ -785,25 +310,26 @@ iso15118::session::feedback::Callbacks ISO15118_chargerImpl::create_callbacks() 
             }
 
             if (control_mode == dt::ControlMode::Scheduled) {
-                charging_needs.control_mode = ControlMode::ScheduledControl;
+                charging_needs.control_mode = types::iso15118::ControlMode::ScheduledControl;
             } else if (control_mode == dt::ControlMode::Dynamic) {
-                charging_needs.control_mode = ControlMode::DynamicControl;
+                charging_needs.control_mode = types::iso15118::ControlMode::DynamicControl;
             } else {
                 EVLOG_error << "Invalid value received for control mode! Not sending 'ChargingNeeds'.";
                 return;
             }
 
             if (mobility_needs_mode == dt::MobilityNeedsMode::ProvidedByEvcc) {
-                charging_needs.mobility_needs_mode = MobilityNeedsMode::EVCC;
+                charging_needs.mobility_needs_mode = types::iso15118::MobilityNeedsMode::EVCC;
             } else if (mobility_needs_mode == dt::MobilityNeedsMode::ProvidedBySecc) {
-                charging_needs.mobility_needs_mode = MobilityNeedsMode::EVCC_SECC;
+                charging_needs.mobility_needs_mode = types::iso15118::MobilityNeedsMode::EVCC_SECC;
             } else {
                 EVLOG_error << "Invalid value received for mobility needs mode! Not sending 'ChargingNeeds'.";
                 return;
             }
 
             // For dash20 the data we will publish will be the v2xChargingParameters
-            V2XChargingParameters& v2x_charging_parameters = charging_needs.v2x_charging_parameters.emplace();
+            types::iso15118::V2XChargingParameters& v2x_charging_parameters =
+                charging_needs.v2x_charging_parameters.emplace();
 
             if (const auto* dc_evse_limits = std::get_if<iso15118::d20::DcTransferLimits>(&evse_limits)) {
                 if (const auto* dc_ev_limits = std::get_if<dt::DC_CPDReqEnergyTransferMode>(&ev_limits)) {
